# chapter10/java/problem-reports-full/src/main/resources/static/swagger.yaml
swagger: "2.0"
info:
  description: "This is the sample service for Part II of the book \"Creating Maintainable\
    \ APIs.\" The service exposes a level 3 HTTP REST API (hypermedia-driven) for\
    \ storing and searching problem reports (per RFC 7807)."
  version: "0.2.0"
  title: "Problem Reports Service"
  contact:
    name: Ervin Varga
    email: e.varga@ieee.org
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0"
host: "localhost:8080"
basePath: "/api/ProblemReports"
...
schemes:
- "http"
consumes:
- "application/json"
- "text/xml"
- "application/xml"
produces:
- "application/json"
- "text/xml"
- "application/xml"
- "application/atom+xml"
- "application/hal+json"
...
paths:
...
  /Reports/{reportNumber}:
    parameters:
    - $ref: "#/parameters/Accept"
    - name: "reportNumber"
      description: "The unique report number."
      in: "path"
      type: "string"
      required: true
    get:
      security:
      - problem-reports-realm: []
      tags:
      - "Reports"
      summary: "Gets data about an individual problem report."
      responses:
        200:
          description: "The data associated with this problem report, and links to\
            \ available options regarding this entity."
          schema:
            $ref: "#/definitions/ProblemReport"
        default:
          $ref: "#/responses/UnexpectedServiceError"
    delete:
      security:
      - problem-reports-realm: []
      tags:
      - "Reports"
      summary: "Deletes the given problem report."
      responses:
        204:
          description: "This problem report has been removed from the system."
        default:
          $ref: "#/responses/UnexpectedServiceError"
...
securityDefinitions:
  problem-reports-realm:
    description: "The service should be protected by a minimal basic authentication\
      \ mechanism to prevent inadvertent usage (the assumption is that this service\
      \ will run behind a corporate firewall)."
    type: "basic"
parameters:
  Accept:
    name: "Accept"
    in: "header"
    description: "The media type requested by a client: application/json,\
      \ application/xml, application/hal+json or application/atom+xml."
    required: false
    default: "application/json"
    type: "string"
    x-example: "Accept: application/atom+xml"
responses:
  UnexpectedServiceError:
    description: Unexpected service error (4x or 5x type of error).
    schema:
      $ref: "#/definitions/ProblemReport"
definitions:
  Links:
    description: "Contains links related to all resource representations."
    type: "array"
    minItems: 1
    uniqueItems: true    
    items:
      type: "object"
      title: "LinkItem"
      properties:
        href:
          type: "string"
          description: "The URL to the resource representation."
        rel:
          type: "string"
          description: "The relation's name; for example, self, about, contents, etc."
      required:
      - "href"
      - "rel"
...
  ProblemReport:
    type: "object"
    required:
    - "type"
    - "title"
    - "status"
    - "correlationId"
    properties:
      links:
        $ref: "#/definitions/Links"
      reportNumber:
        type: "string"
        description: "The globally unique report identifier."
      type:
        type: "string"
        description: "The type of this report (used to categorize problems)."
      title:
        type: "string"
        description: "The name of this problem report."
      status:
        type: "integer"
        description: "The status code associated with this problem."
      correlationId:
        type: "string"
        description: "The identifier to group related reports together."
    description: "The resource representation of an individual problem report.\ 
      \ It contains dynamic links to drive the behavior via hypermedia controls."
...